Refactor dependency cycle
